requirements.txt
Flask==2.0.3
slackclient==2.9.3
slackeventsapi==3.0.1
gunicorn==20.1.0
